添加武器消耗逻辑

+++ Imitate-Soul-Knight-Project/Assets/Scripts/Data/Config/WeaponConfigData.cs
@@ -14,7 +14,7 @@ public class WeaponConfigData : BaseData {
 	public int launchCount; // 发射数量
 	public int bulletSpeed; // 子弹速度
 	public string bulletUrl; // 子弹路径
-	public float mpConsume; // 蓝量消耗值
+	public int mpConsume; // 蓝量消耗值
 	public float bulletOffset; // 子弹偏移度
 	public float recoilForceDis; // 后坐力作用距离
 }
